BtSerial: register GAP callback in master and slave modes
Currently, the GAP callback is registered only if the device is in
master mode. This is incorrect, since the callback is required also
as a slave in order to perform GAP related tasks, like confirming
the SSP pairing.

This patch removes the restriction of being master to register the
GAP callback, so it works in both slave and master modes.